Fix: the linter issues from rubocop
