+++ creation.rb
@@ -2,6 +2,7 @@ require 'json'
 require_relative './student'
 require_relative './teacher'
 require_relative './store'
+require_relative './helper'
 
 module Creation
   def self.success(item)
@@ -70,40 +71,12 @@ module Creation
     success('Book')
   end
 
-  def self.books_and_people_to_rental
-    if(File.exists?("books.json")) 
-      books_arr = JSON.parse(File.read('books.json'))
-    else
-      File.open("books.json","w")
-      File.write("books.json", "[]")
-    end
-    puts 'Select a book from the following list by number'
-    books_arr.each_with_index { |book, idx| puts "#{idx}) Title: \"#{book['title']}\" Author: #{book['author']}" }
-    book_index = gets.chomp.to_i
-    puts 'Select a person from the following list by number (not id)'
-    if(File.exists?("people.json")) 
-      people_arr = JSON.parse(File.read('people.json'))
-    else
-      File.open("people.json","w")
-      File.write("people.json", "[]")
-    end
-    people_arr.each_with_index do |person, idx|
-      puts "#{idx}) [#{person['type']}] Name: #{person['name']}, ID: #{person['id']}, Age: #{person['age']}"
-    end
-    person_index = gets.chomp.to_i
-    [books_arr, people_arr, book_index, person_index]
-  end
-
   def self.create_a_rental
-    books_arr, people_arr, book_index, person_index = books_and_people_to_rental
+    books_arr, people_arr, book_index, person_index = Store.books_and_people_to_rental
     print 'Date (yyyy/mm/dd): '
     date = gets.chomp
-    if(File.exists?("rentals.json")) 
-      rentals_arr = JSON.parse(File.read('rentals.json'))
-    else
-      File.open("books.json","w")
-      File.write("books.json", "[]")
-    end
+    Helper.create_file_if_not_exist('rentals.json')
+    rentals_arr = JSON.parse(File.read('rentals.json'))
     person = people_arr[person_index]
     book = books_arr[book_index]
     rental = Rental.new(date, person, book)
@@ -114,8 +87,7 @@ module Creation
       end
     end
     File.write('people.json', JSON.generate(new_people_arr))
-    json_to_arr = rentals_arr.push({ 'date' => rental.date, 'person' => rental.person,
-                                     'book' => rental.book })
+    json_to_arr = rentals_arr.push({ 'date' => rental.date, 'person' => rental.person, 'book' => rental.book })
     File.write('rentals.json', JSON.generate(json_to_arr))
     success('Rental')
   end
