+++ store.rb
@@ -1,4 +1,5 @@
 require 'json'
+require_relative './helper'
 
 class Store
   def self.read_convert(filename)
@@ -7,13 +8,8 @@ class Store
 
   def self.push(item)
     if item.is_a?(Person)
-
-      if(File.exists?("people.json") == false)
-        File.open("people.json","w")
-        File.write("people.json", "[]")
-      end
+      Helper.create_file_if_not_exist('people.json')
       json_to_arr = read_convert('people.json')
-
 
       if item.type == 'Student'
         json_to_arr.push({ type: item.type, 'name' => item.name, 'id' => item.id, 'age' => item.age,
@@ -24,32 +20,29 @@ class Store
       end
       File.write('people.json', JSON.generate(json_to_arr))
     else
-      if(File.exists?("books.json") == false)
-        File.open("books.json","w")
-        File.write("books.json", "[]")
-      end
+      Helper.create_file_if_not_exist('books.json')
       json_to_arr = read_convert('books.json').push({ 'title' => item.title, 'author' => item.author })
       File.write('books.json', JSON.generate(json_to_arr))
     end
   end
 
   def self.list_all_books
-    if(File.exists?("books.json")) 
+    if File.exist?('books.json')
       JSON.parse(File.read('books.json')).map { |book| puts "Title: \"#{book['title']}\" Author: #{book['author']}" }
     else
-      File.open("books.json","w")
-      File.write("books.json", "[]")
+      File.open('books.json', 'w')
+      File.write('books.json', '[]')
     end
   end
 
   def self.list_all_people
-    if(File.exists?("people.json")) 
+    if File.exist?('people.json')
       JSON.parse(File.read('people.json')).map do |person|
         puts "[#{person['type']}] Name: #{person['name']}, ID: #{person['id']}, Age: #{person['age']}"
       end
     else
-      File.open("people.json","w")
-      File.write("people.json", "[]")
+      File.open('people.json', 'w')
+      File.write('people.json', '[]')
     end
   end
 
@@ -68,5 +61,25 @@ class Store
     collection = []
     ObjectSpace.each_object(Store) { |item| collection << item }
     collection
+  end
+
+  def self.books_and_people_to_rental
+    Helper.create_file_if_not_exist('books.json')
+    books_arr = JSON.parse(File.read('books.json'))
+    puts 'Select a book from the following list by number'
+    books_arr.each_with_index { |book, idx| puts "#{idx}) Title: \"#{book['title']}\" Author: #{book['author']}" }
+    book_index = gets.chomp.to_i
+    puts 'Select a person from the following list by number (not id)'
+    if File.exist?('people.json')
+      people_arr = JSON.parse(File.read('people.json'))
+    else
+      File.open('people.json', 'w')
+      File.write('people.json', '[]')
+    end
+    people_arr.each_with_index do |person, idx|
+      puts "#{idx}) [#{person['type']}] Name: #{person['name']}, ID: #{person['id']}, Age: #{person['age']}"
+    end
+    person_index = gets.chomp.to_i
+    [books_arr, people_arr, book_index, person_index]
   end
 end
